+++ clients/dns-service-client/src/lib.rs
@@ -26,12 +26,11 @@ progenitor::generate_api!(
     }
 );
 
-pub type DnsError = crate::Error<crate::types::Error>;
+pub use internal_dns_types::config::{
+    ERROR_CODE_BAD_UPDATE_GENERATION, ERROR_CODE_UPDATE_IN_PROGRESS,
+};
 
-pub const ERROR_CODE_UPDATE_IN_PROGRESS: &'static str = "UpdateInProgress";
-pub const ERROR_CODE_BAD_UPDATE_GENERATION: &'static str =
-    "BadUpdateGeneration";
-pub const ERROR_CODE_INCOMPATIBLE_RECORD: &'static str = "IncompatibleRecord";
+pub type DnsError = crate::Error<crate::types::Error>;
 
 /// Returns whether an error from this client should be retried
 pub fn is_retryable(error: &DnsError) -> bool {
