+++ dns-server-api/Cargo.toml
@@ -12,6 +12,7 @@ chrono.workspace = true
 dropshot.workspace = true
 dropshot-api-manager-types.workspace = true
 internal-dns-types.workspace = true
+internal-dns-types-versions.workspace = true
 omicron-workspace-hack.workspace = true
 schemars.workspace = true
 semver.workspace = true
