+++ dns-server/src/http_server.rs
@@ -6,15 +6,11 @@
 
 use crate::storage::{self, UpdateError};
 use dns_server_api::DnsServerApi;
-use dns_service_client::{
-    ERROR_CODE_BAD_UPDATE_GENERATION, ERROR_CODE_INCOMPATIBLE_RECORD,
-    ERROR_CODE_UPDATE_IN_PROGRESS,
-};
 use dropshot::RequestContext;
-use internal_dns_types::{
-    v1::{self, config::TranslationError as V1TranslationError},
-    v2::{self, config::TranslationError as V2TranslationError},
+use internal_dns_types::config::{
+    ERROR_CODE_BAD_UPDATE_GENERATION, ERROR_CODE_UPDATE_IN_PROGRESS,
 };
+use internal_dns_types_versions::latest;
 
 pub struct Context {
     store: storage::Store,
@@ -36,64 +32,10 @@ enum DnsServerApiImpl {}
 impl DnsServerApi for DnsServerApiImpl {
     type Context = Context;
 
-    async fn dns_config_get_v1(
-        rqctx: RequestContext<Context>,
-    ) -> Result<
-        dropshot::HttpResponseOk<v1::config::DnsConfig>,
-        dropshot::HttpError,
-    > {
-        let result = Self::dns_config_get(rqctx).await?;
-        match result.0.try_into() {
-            Ok(config) => Ok(dropshot::HttpResponseOk(config)),
-            Err(V2TranslationError::IncompatibleRecord) => {
-                Err(dropshot::HttpError::for_bad_request(
-                    None,
-                    ERROR_CODE_INCOMPATIBLE_RECORD.to_string(),
-                ))
-            }
-        }
-    }
-
-    async fn dns_config_get_v2(
-        rqctx: RequestContext<Context>,
-    ) -> Result<
-        dropshot::HttpResponseOk<v2::config::DnsConfig>,
-        dropshot::HttpError,
-    > {
-        Self::dns_config_get(rqctx).await
-    }
-
-    async fn dns_config_put_v1(
-        rqctx: RequestContext<Context>,
-        rq: dropshot::TypedBody<v1::config::DnsConfigParams>,
-    ) -> Result<dropshot::HttpResponseUpdatedNoContent, dropshot::HttpError>
-    {
-        let provided_config = match rq.into_inner().try_into() {
-            Ok(config) => config,
-            Err(V1TranslationError::GenerationTooLarge) => {
-                return Err(dropshot::HttpError::for_bad_request(
-                    None,
-                    ERROR_CODE_INCOMPATIBLE_RECORD.to_string(),
-                ));
-            }
-        };
-        Self::dns_config_put(rqctx, provided_config).await
-    }
-
-    async fn dns_config_put_v2(
-        rqctx: RequestContext<Context>,
-        rq: dropshot::TypedBody<v2::config::DnsConfigParams>,
-    ) -> Result<dropshot::HttpResponseUpdatedNoContent, dropshot::HttpError>
-    {
-        Self::dns_config_put(rqctx, rq.into_inner()).await
-    }
-}
-
-impl DnsServerApiImpl {
     async fn dns_config_get(
         rqctx: RequestContext<Context>,
     ) -> Result<
-        dropshot::HttpResponseOk<v2::config::DnsConfig>,
+        dropshot::HttpResponseOk<latest::config::DnsConfig>,
         dropshot::HttpError,
     > {
         let apictx = rqctx.context();
@@ -108,11 +50,14 @@ impl DnsServerApiImpl {
 
     async fn dns_config_put(
         rqctx: RequestContext<Context>,
-        params: v2::config::DnsConfigParams,
+        rq: dropshot::TypedBody<latest::config::DnsConfigParams>,
     ) -> Result<dropshot::HttpResponseUpdatedNoContent, dropshot::HttpError>
     {
         let apictx = rqctx.context();
-        apictx.store.dns_config_update(&params, &rqctx.request_id).await?;
+        apictx
+            .store
+            .dns_config_update(&rq.into_inner(), &rqctx.request_id)
+            .await?;
         Ok(dropshot::HttpResponseUpdatedNoContent())
     }
 }
