+++ internal-dns/types/versions/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "internal-dns-types"
+name = "internal-dns-types-versions"
 version = "0.1.0"
 edition.workspace = true
 
@@ -9,14 +9,7 @@ workspace = true
 [dependencies]
 anyhow.workspace = true
 chrono.workspace = true
-internal-dns-types-versions.workspace = true
 omicron-common.workspace = true
 omicron-workspace-hack.workspace = true
-omicron-uuid-kinds.workspace = true
 schemars.workspace = true
 serde.workspace = true
-strum.workspace = true
-
-[dev-dependencies]
-expectorate.workspace = true
-serde_json.workspace = true
