+++ internal-dns/types/versions/src/initial/mod.rs
@@ -2,4 +2,6 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
+//! DNS configuration types for API version 1.0.0 (INITIAL).
+
 pub mod config;
