+++ internal-dns/types/versions/src/latest.rs
@@ -1,0 +1,13 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at https://mozilla.org/MPL/2.0/.
+
+//! Re-exports of the latest versions of each type.
+
+pub mod config {
+    pub use crate::v2::config::DnsConfig;
+    pub use crate::v2::config::DnsConfigParams;
+    pub use crate::v2::config::DnsConfigZone;
+    pub use crate::v2::config::DnsRecord;
+    pub use crate::v2::config::Srv;
+}
